+++ app/controllers/v1/categories_controller.rb
