+++ app/controllers/v1/entries_controller.rb
